classfication/BAG/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.2.0
openai==1.12.0
python-dotenv==1.0.1
transformers>=4.11.0
numpy>=1.19.5
scikit-learn>=0.24.2
tensorboard>=2.7.0
lxml==5.1.0
selenium==4.18.1
webdriver-manager==4.0.1
undetected-chromedriver==3.5.5
seleniumbase==4.22.2
torch>=1.10.0
torchvision>=0.11.0
tqdm>=4.62.3
matplotlib>=3.5.0
networkx>=2.6.3
urllib3>=1.26.7
torch_scatter==2.1.2
Pillow>=8.0.0